Package files into corresponding packages
