--- data/Saitama.txt
@@ -1,2 +1,0 @@
-T 0 eat
-T 0 eat
