+++ src/main/java/saitama/Ui.java
@@ -1,3 +1,5 @@
+package saitama;
+
 import java.util.Scanner;
 
 public class Ui {
