+++ src/main/java/saitama/commands/ExitCommand.java
@@ -1,4 +1,10 @@
-class ExitCommand extends Command {
+package saitama.commands;
+
+import saitama.Storage;
+import saitama.TaskList;
+import saitama.Ui;
+
+public class ExitCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         ui.showSave();
