+++ src/main/java/saitama/exceptions/EmptyDescriptionException.java
@@ -1,5 +1,7 @@
+package  saitama.exceptions;
+
 public class EmptyDescriptionException extends SaitamaException {
-    EmptyDescriptionException() {
+    public EmptyDescriptionException() {
         super("OOPS!!! The task description cannot be empty.");
     }
 }
