+++ src/main/java/saitama/tasks/Deadline.java
@@ -1,3 +1,7 @@
+package saitama.tasks;
+
+import saitama.exceptions.InvalidFormatException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
