+++ src/main/java/saitama/tasks/Event.java
@@ -1,3 +1,5 @@
+package saitama.tasks;
+
 public class Event extends Task {
 
     protected String at;
