+++ src/main/java/saitama/tasks/Task.java
@@ -1,3 +1,5 @@
+package saitama.tasks;
+
 import java.io.IOException;
 
 abstract public class Task {
